Fix mouse coordinate system and update dependencies

+++ bracket-random/src/random.rs
@@ -26,7 +26,7 @@ fn get_seed() -> u64 {
         SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .unwrap()
-            .as_secs() as u64
+            .as_secs()
     }
 }
